Query timeout now properly set for each query
As per Apple’s docs, changing the configuration of an `NSURLSession` instance after its initialization has no effect. As a consequence, the timeout failed to be taken into account. Now, the timeout is explicitly set for each query.

I took the opportunity to set a different (longer) timeout for data selection queries when synchronizing the offline mirror.