[offline] Preventive offline search requests
When an index is mirrored locally, for every online search request, an offline search request is launched preventively after a certain delay. If the online request hasn’t finished before the offline request finishes, the completion handler is called with the offline results. When the online request finishes successfully, the completion handler will be called a second time.