Implement missing `OfflineIndex` features
- Delete by query: custom implementation
- Disjunctive faceting: factorized with `Index`
- Multiple queries: factorized with `MirroredIndex`

Also added a pseudo-unit-test for `OfflineIndex`. Because of completely insane Cocoapods issues, unit tests cannot work when we need an external dependency. => I have built a temporary iOS app to run some tests. It’s better than nothing… :/ (And maybe one day, Cocoapods will be fixed and we will be able to run proper unit tests on the offline mode.)