Follow Swift 3’s convention to use lower camel case for enums
It required a twist for the `TypoTolerance` type, which has values “true” and “false”. The enum has a case `bool` with an associated `Bool` value.

There is also a collision between the `none` value that some of our enums have and `Optional.none`. Since `Query`’s properties are all optionals, assigning `.none` when using type inference is actually equivalent to assigning `nil`… :/