Use `Error` instead of `NSError`
`Error` is the new, official protocol for error handling in Swift 3, so let’s use it. `URLSession` uses it already anyway.

This allows us to use different classes, a bit like exceptions in other languages.

`Error` is bridged to `NSError` in Objective-C, but we need some custom code to propagate the error code.