Improve type safety in the `Query` class (HUGE)
For value-typed properties that are not directly bridgeable to Objective-C, the logic has been reversed compared to version 3.x:

- The strongly typed, Swift-only property is named after the parameter.
- The weakly typed, Objective-C bridged property is named with a trailing underscore… but mapped without the underscore in Objective-C.

This way, each language sees the most appropriate type. The only drawback is that Swift also sees the underscore-suffixed properties. (This was already the case for enums.)

Also switched the dual-typed properties like `removeStopWords` (boolean or string array) to an enum with associated values. (This was already the case for `typoTolerance`.)