Refactor `Query` class for more convenient properties (HUGE)
The Swift class is no longer polluted by Objective-C specific properties. Those have been moved to a subclass `_objc_Query`, which is the one that is exposed to Objective-C as `Query`.

It involves quite a lot of juggling between types, properties and Objective-C names, but in the end it should make for a better interface both in Swift and Objective-C.