[refact] Switch back to only one `Query` class
The two-class trick was elegant, but:

1. It makes it impossible to re-use the `Query` class in other libraries, either by aggregation or inheritance.

2. It exposes the `BaseQuery` type in method signatures in Objective-C.

For that reason, I backtracked to using only one class. The Objective-C specific properties are named with a `z_` prefix, which should prevent them from appearing first in autocompletion, and undocumented, which should prevent them to appear in the reference documentation.

NOTE: This change is backward-compatible at the source level, provided that:

- Objective-C code did not reference the `BaseQuery` type explicitly
- Swift code did not reference the `_objc_Query` type explicitly