[offline] Use network reachability to decide whether to perform online requests (#168)
* [offline] Use network reachability to decide whether to perform online requests

The system’s behavior has apparently changed. Network requests used to fail fast when the device was offline. Now, they just hit the timeout, which of course is an entirely different story. Therefore we use `SCNetworkReachability` to test network connectivity and launch online requests based on the result.

* Use network reachability also in the online client

All online requests will now fail fast if reachability says that the device is offline, unless the developer disables this behavior by setting `AbstractClient.useReachability` to `false`.

NOTE: As this feature uses the System Configuration framework, it is not available on watchOS. The behavior remains unchanged on that platform.