New offline features (#185)
* [offline] Support bootstrapping mirrored indices with local data

This feature allows initializing a local mirror from local data (typically, prepackaged resources) without having to wait for the first sync.

* [offline] Expose `hasOfflineData()` at the index level

* [offline] Add a `buildOffline(…)` method on `MirroredIndex`

This unconditionally creates or replaces the local mirror with the specified data, regardless of whether local data already exists or not for the index. (In contrast to `boostrap(…)`, which only does so if no data exists.)

* [offline] Support offline fallback for getting objects (WIP)

The `Index.getObjects(…)` methods are overridden by `MirroredIndex`, providing the same offline fallback mechanism as search.

NOTE: Only supported for `getObjects` so far; getting a single object will follow later.

* [offline] Fix race condition in instantiation of local index

It turns out that [lazy variables are not thread-safe](http://stackoverflow.com/a/29761777). We therefore have to synchronize initialization manually.

* [offline] Refactor recent features

- Remove `bootstrap(…)`: it is superseded by `buildOffline(…)`
- Use a dynamic property instead of a function for `hasOfflineData`
- Rename notifications from `bootstrapDid*` to `buildDid*`
- Update reference documentation

* [offline] Leverage `buildOffline` inside sync

This allows to send build notifications also for a sync.

* [offline] Add unit test for `MirroredIndex.buildOffline(…)`

* Fix compilation warnings

The `syncErrorKey` notification key is now deprecated in favor of `errorKey`.

* [offline] Add utilities to tag content origin as local or remote

* [offline] Add offline support to `MirroredIndex.getObject(…)`

* [offline] Fix ambiguous overload

* [refact] Use utilities to tag content origin

* Ooops

* [offline] Add test cases for `MirroredIndex.getObject[s]`

* [offline] Fix compilation of unit tests

Xcode or Cocoapods (not sure which one) has trouble dealing with two distinct modules both named “AlgoliaSearch”. We therefore rename the offline flavor to “AlgoliaSearchOffline”.

NOTE: This is **not** what we want clients of the library to be using, but since the module name is overridden by Cocoapods, and the scheme is not shared with Carthage, this should impact only the unit tests, not the users of the module.

* [refact] Add completion handler to `MirroredIndex.buildOffline(…)`

* [refact] Use dynamic property instead of function for `OfflineIndex.hasOfflineData`

* [offline] Add manual build capability to `OfflineIndex`

* Fix some inconsistencies

* [doc][offline] Update reference documentation

[ci skip]

* [doc][offline] Fix generation of reference documentation

Changing the module name had unintended side effects on the documentation. This should fix it.

* [refact][offline] Return an operation from manual build methods

* [offline] Assert that mirroring is on before testing local data

* [offline][doc] Update reference documentation

[ci skip]