Fix race condition when searching while activating mirrored mode
Although instantiation of `MirroredIndex.localIndex` was conditioned to setting `MirroredIndex.mirrored` to true, and although `mirrored`’s accessors were synchronized, the instantiation itself was not part of the critical section. This led to a race condition when `search(…)` is called from another thread after `mirrored` has been set but before `localIndex` has been instantiated.

NOTE: I had to give up deallocating the local index when setting `mirrored` back to false, because mixed operations or a sync may still be ongoing. Keeping it alive until all operations have completed would be extremely complicated and not worth it, because there is anyway little point in disabling mirrored mode once it has been activated. This flag is merely so that purely online indices can still be used with an `OfflineClient`.

Fixes #285.