[refact] `User-Agent` header now a static property
This is required to abstract InstantSearch Core’s `Searcher` class from the `Index` class. Otherwise we cannot access the `Client` instance, therefore we cannot patch the user agent. By making it a global property, we allow patching from anywhere.

There is now an `addUserAgent(…)` method that takes care of deduplication.

Patching of headers is now done at the `Request` level, which means that users cannot accidentally erase the `User-Agent` header: they can only add to it.

WARNING: Small backward-compatibility break since the instance property is no longer writable. But it was documented as for “internal purposes only”, so nobody should use it.